Feat: 모바일 내비게이션 재구성 — 다음세대 대신 교회 소식 진입, /about 직접 렌더, news 형제 탭 복원
- 왜: 실기기 QA에서 하단 내비게이션이 미구현 '다음세대'로 들어가고, /about은 redirect로만 열리고, 교회 소식 목록에 형제 탭이 안 붙는 문제를 발견했다.
- 무엇:
  - BottomNav(items 2·4): IconName에 news 추가·nextgen 제거, ICON_MAP에 news → LuNewspaper 매핑, BOTTOM_NAV_ITEMS를 홈 > 교회 소개 > 설교 > 교회 소식 > 마이페이지로 재정렬(다음세대 제거).
  - /about(item 3): redirect를 없애고 page 단에서 PastorGreeting을 직접 렌더하고 metadata를 붙인다. 인사말 JSX와 getPastorPageData를 about/_component/PastorGreeting으로 추출해 /about·/about/pastor가 공유하고, canonical은 둘 다 /about/pastor로 맞춘다. AboutTabNav의 인사말 탭을 두 경로 모두에서 활성으로 표시한다.
  - 교회 소식 탭(item 5): NEWS_TABS(주보 > 공지사항 > 갤러리)와 isNewsListPath를 신설한다. resolveSiblingTabs가 목록 경로에만 형제 탭을, resolveMobileHeader가 '교회 소식' 헤더를 반환한다. /news는 /news/bulletins로 redirect한다 — 재export는 활성 탭 계산에서 어느 탭과도 안 맞아 헤더·탭이 안 붙었다.
- 영향: 뉴스 상세(/news/[id])는 현행 헤더·공유·뒤로가기를 유지하고 형제 탭은 목록에만 붙는다. GNB children·드로어·데스크톱 드롭다운 순서는 그대로 둔다 — news 전용 배열로 분리했다. exec-plan 신설(docs/exec-plans/active/2026-07-04-mobile-nav-qa.md).
- 제외: 다음세대(/next-gen) 페이지 자체는 안 건드린다 — 하단 내비게이션 진입만 교체.
