+++ src/app/(content)/about/_component/PastorGreeting.module.scss
